requirements.txt
# Core deep learning libraries
torch>=1.12.0
torchvision>=0.13.0
pytorch-lightning>=1.8.0

# Computer vision và xử lý video
opencv-python>=4.6.0
Pillow>=9.0.0
opencv-contrib-python>=4.6.0

# GUI framework
tkinter  # Built-in với Python, không cần cài

# Web framework (cho app.py web version)
Flask>=2.2.0
Werkzeug>=2.2.0

# Data science và visualization
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0
pandas>=1.4.0

# Explainable AI
grad-cam>=1.4.0

# Video processing utilities
imageio>=2.20.0
imageio-ffmpeg>=0.4.7

# Progress bars và utilities
tqdm>=4.64.0
requests>=2.28.0

# JSON và file handling
jsonschema>=4.0.0

# Threading và multiprocessing
concurrent-futures>=3.1.1

# Cross-platform video player support
playsound>=1.3.0  # Cho audio
pygame>=2.1.0     # Cho multimedia playback

# Optional: Để tăng hiệu suất
# accelerate
# xformers
